split off fetch.
