+++ src/hooks/index.ts
@@ -13,4 +13,3 @@ export { default as useVoteTiming } from "./useVoteTiming";
 export { default as useCurrentRoundId } from "./useCurrentRoundId";
 export { default as useRound } from "./useRound";
 export { default as useMulticall } from "./useMulticall";
-export { default as useUMIP } from "./useUMIP";
